Move utils (#5)
* separate `utils` into files and fix imports

* last import fix
