+++ extensions/development/db/ext.py
@@ -4,7 +4,7 @@ __all__ = ("DB",)
 from interactions.ext.prefixed_commands.manager import PrefixedInjectedClient
 
 from AlbertoX3.ipy_wrapper import Extension
-from AlbertoX3.utils import get_logger
+from AlbertoX3.utils.essentials import get_logger
 
 
 logger = get_logger()
